requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
duckduckgo_search==7.5.1
exceptiongroup==1.2.2
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lancedb==0.14.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.25.2
numpy==1.26.4
openai==1.51.2
overrides==7.7.0
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pgvector==0.3.5
phidata==2.5.5
pillow==11.1.0
platformdirs==4.3.6
#primp==0.12.0
primp==0.14.0
protobuf==5.29.3
psycopg==3.2.3
py==1.11.0
pyarrow==19.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pylance==0.18.2
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retry==0.9.2
rich==13.9.2
rpds-py==0.22.3
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
streamlit==1.42.0
tantivy==0.22.0
tenacity==9.0.0
toml==0.10.2
tomli==2.0.2
tornado==6.4.2
tqdm==4.67.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
webencodings==0.5.1
wheel==0.45.1
yfinance==0.2.44
firecrawl-py
#googlesearch
#pycountry

